Small bugfixes and some slight refactoring
It was possible "pause" the game in the MenuState.
Fixed that by adding an extra check. However, it should be noted
that this bug only occured because the player is already created even
though the game has not been started yet.

Added releaseAllKeys() inside player object, so in case the user pauses
the game while moving/shooting/etc, the keys do not get stuck.

Removed unused method, and some renaming (removed "get" prefix")